phase3-logic completed
buy,rate,delete,comments work fine.
filter movies has bug!

+++ handlers.cpp
@@ -3,25 +3,29 @@
 
 using namespace std;
 
+Response *HomeHandler::callback(Request *req) {
+  int user_id = stoi(req->getSessionId());
+  Response *res = new Response();
+  if(UT_flix->get_user_type(user_id) == USER_PUBLISHER)
+    res = Response::redirect("/publisherhome");
+  else
+    res = Response::redirect("/userhome");
+  return res;
+}
 Response *LoginHandler::callback(Request *req) {
   string username = req->getBodyParam("username");
   string password = req->getBodyParam("password");
   string id = to_string(UT_flix->login(username,password));
   Response *res = new Response;
   res->setSessionId(id);
-  if(UT_flix->get_user_type(stoi(id)) == USER_PUBLISHER)
-    res = Response::redirect("/publisherhome");
-  else
-    res = Response::redirect("/userhome");
+  res = Response::redirect("/home");
   return res;
 }
+
 Response *SignupHandler::callback(Request *req) {
   string id = to_string(UT_flix->signup(req));
   Response *res = new Response;
-  if(UT_flix->get_user_type(stoi(id)) == USER_PUBLISHER)
-    res = Response::redirect("/publisherhome");
-  else
-    res = Response::redirect("/userhome");
+  res = Response::redirect("/home");
   res->setSessionId(id);
   return res;
 }
@@ -32,12 +36,20 @@ Response *LogoutHandler::callback(Request *req) {
   return res;
 }
 Response *PublishFilmsHandler::callback(Request *req) {
+  if(req->getSessionId() == DEFAULT_SESSION_ID){
+    Response *res = Response::redirect("/error_page");
+    return res;
+  }
   int user_id = stoi(req->getSessionId());
   UT_flix->upload_films(req,user_id);
   Response *res = Response::redirect("/publisherhome");
   return res;
 }
 Response *DeleteFilmsHandler::callback(Request *req) {
+  if(req->getSessionId() == DEFAULT_SESSION_ID){
+    Response *res = Response::redirect("/error_page");
+    return res;
+  }
   string film_id = req->getQueryParam("film_id");
   int user_id = stoi(req->getSessionId());
   UT_flix->delete_movie(film_id, user_id);
@@ -45,14 +57,22 @@ Response *DeleteFilmsHandler::callback(Request *req) {
   return res;
 }
 Response *RateFilmHandler::callback(Request *req) {
-  string film_id = req->getBodyParam("film_id");
-  string score = req->getBodyParam("score");
+  if(req->getSessionId() == DEFAULT_SESSION_ID){
+    Response *res = Response::redirect("/error_page");
+    return res;
+  }
+  string film_id = req->getQueryParam("film_id");
+  string score = req->getQueryParam("score");
   int user_id = stoi(req->getSessionId());
   UT_flix->rate_movie(film_id, score, user_id);
   Response *res = Response::redirect("/userhome");
   return res;
 }
 Response *BuyFilmHandler::callback(Request *req) {
+  if(req->getSessionId() == DEFAULT_SESSION_ID){
+    Response *res = Response::redirect("/error_page");
+    return res;
+  }
   string film_id = req->getQueryParam("film_id");
   int user_id = stoi(req->getSessionId());
   UT_flix->buy_movie(film_id, user_id);
@@ -60,6 +80,10 @@ Response *BuyFilmHandler::callback(Request *req) {
   return res;
 }
 Response *FilmDetailsHandler::callback(Request *req) {
+  if(req->getSessionId() == DEFAULT_SESSION_ID){
+    Response *res = Response::redirect("/error_page");
+    return res;
+  }
   int film_id = stoi(req->getQueryParam("film_id"));
   int user_id = stoi(req->getSessionId());
   Response *res = new Response;
@@ -69,6 +93,10 @@ Response *FilmDetailsHandler::callback(Request *req) {
 
 }
 Response *IncreaseMoneyHandler::callback(Request *req) {
+  if(req->getSessionId() == DEFAULT_SESSION_ID){
+    Response *res = Response::redirect("/error_page");
+    return res;
+  }
   std::string amount = req->getQueryParam("amount");
   int user_id = stoi(req->getSessionId());
   UT_flix->increase_user_money(amount, user_id);
@@ -77,6 +105,10 @@ Response *IncreaseMoneyHandler::callback(Request *req) {
 }
 
 Response *PublisherHandler::callback(Request *req) {
+  if(req->getSessionId() == DEFAULT_SESSION_ID){
+    Response *res = Response::redirect("/error_page");
+    return res;
+  }
   int user_id = stoi(req->getSessionId());
   Response *res = new Response();
   res->setHeader("Content-Type","text/html");
@@ -84,16 +116,45 @@ Response *PublisherHandler::callback(Request *req) {
   return res;
 }
 Response *UserHandler::callback(Request *req) {
+  if(req->getSessionId() == DEFAULT_SESSION_ID){
+    Response *res = Response::redirect("/error_page");
+    return res;
+  }
   int user_id = stoi(req->getSessionId());
   Response *res = new Response();
   res->setHeader("Content-Type","text/html");
-  res->setBody(UT_flix->user_views_available_movies(user_id));
+  res->setBody(UT_flix->user_views_available_movies(req,user_id));
   return res;
 }
 Response *ProfileHandler::callback(Request *req) {
+  if(req->getSessionId() == DEFAULT_SESSION_ID){
+    Response *res = Response::redirect("/error_page");
+    return res;
+  }
   int user_id = stoi(req->getSessionId());
   Response *res = new Response();
   res->setHeader("Content-Type","text/html");
-  res->setBody(UT_flix->user_views_bought_movies(user_id));
+  res->setBody(UT_flix->user_views_bought_movies(req,user_id));
+  return res;
+}
+Response *CommentHandler::callback(Request *req) {
+  if(req->getSessionId() == DEFAULT_SESSION_ID){
+    Response *res = Response::redirect("/error_page");
+    return res;
+  }
+  int film_id = stoi(req->getQueryParam("film_id"));
+  string content = req->getQueryParam("content");
+  int user_id = stoi(req->getSessionId());
+  UT_flix->comment_on_films(user_id,film_id, content);
+  Response *res = Response::redirect("/userhome");
+  return res;
+}
+
+Response *FilterHandler::callback(Request *req) {
+  if(req->getSessionId() == DEFAULT_SESSION_ID){
+    Response *res = Response::redirect("/error_page");
+    return res;
+  }
+  Response *res = Response::redirect("/home");
   return res;
 }
